Move Round entity to the RoundAggregate
